Moved config module back into core
